refactor: move note related files to '@/components/note'
